Initial commit.
This initial commit is a tiny bit different than the upstream JSLC
as I have already made some small changes in my work on the D3D11
pipeline.

The reason I am creating this repository is because I am not sure how
to run the jslc tests from the upstream repository (they don't seem to
be run with any of the config flags). It is also kind of nice because
the build is really fast. Once I make the necessary changes for JSLC to
support higher pixel and vertex shader versions then I will import the
changes back.